- Fixed segfaults in the C lib due to creation of rocksdb objects during static initialization - Simplified configuration types - Removed experimental GC forcing (in transition_state) in the C lib

+++ taraxa/state/api.go
@@ -30,7 +30,7 @@ type ChainConfig struct {
 }
 type APIOpts struct {
 	// TODO have single "perm-gen size" config property to derive all preallocation sizes
-	ExpectedMaxTrxPerBlock        uint32
+	ExpectedMaxTrxPerBlock        uint64
 	MainTrieFullNodeLevelsToCache byte
 }
